requirements.txt
aiohttp==3.7.4.post0
astroid==2.5.2
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
Cython==0.29.22
discord==1.0.1
discord.py==1.6.0
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
multidict==5.1.0
numpy==1.20.2
pandas==1.2.3
prompt-toolkit==3.0.18
pylint==2.7.4
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
wit==6.0.0
wrapt==1.12.1
yarl==1.6.3
